refactor channel into output
